WIP
pythoncore builds

Separate UWP project tree for simplicity, should be removed and merged into the regular one once experimentation is finished

build fixes for openssl

WIP

build fixes

Some fixes and removal of UWPbuild

More work

fix install of pyconfig.h

Add visualisation of python objects when debugging

Various build fixes and warnings fixes

generate importlib during build

Remove registryfinder as it's not available in uwp and
we don't want to search registry on desktop either

Works!

fix build and exports

fix two unresolved exports
remove asm usage because of weird masm errors that suddenly started

fix win32 build

second attempt

disable safeseh

Fix arm build

Fix arm64 support in pyatomic

downgrade cmake requirements

+++ Modules/_ctypes/stgdict.c
@@ -814,7 +814,7 @@ PyCStructUnionType_update_stgdict(PyObject *type, PyObject *fields, int isStruct
                 element_types[element_index++] = &dict->ffi_type_pointer;
             }
             else {
-                int length = dict->length;
+                int length = (int)dict->length;
                 StgDictObject *edict;
 
                 edict = PyType_stgdict(dict->proto);
